Basic RHCOS publishing structure
A basic easy extensible structure to publish repos.
More generalization is possible, but the basics are there.

The Dockerfile for rhcos always downloads the latest rhco 4.9 image and
verifies that the published shasum matches the download.

The build script on top, makes use of an interface function
`containerdisks::needs_update` which returns the shasum of the new
to-download image, or returns nothing if no update is necessary.

This function can work since every containerdisk gets labeled with the
shasum of the containerdisk file itself. Skopeo queries the remote image
to get the info.

Signed-off-by: Roman Mohr <rmohr@redhat.com>