Append a slash to tar directory entries
CDI detects the disk directory entry as a regular file due to checking
if the tar header points to a directory by checking if it is appended
with a slash instead of checking the tar type.

To make the images works with all all CDI versions as well as meeting
this expectation in general (most apps append a slash), adding a slash
to the directories here too.

Signed-off-by: Roman Mohr <rmohr@redhat.com>