Replace LayerFromReader with LayerFromOpener
The function tarball.LayerFromReader internally uses iotuil.ReadAll, which
buffers all data in memory and can lead to problems with large files.
Replace it with tarball.LayerFromOpener and implement a StreamLayerOpener,
that does not return an error chan and can be opened multiple times.

Thanks to akrejcir for the idea and proposal.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>