Add images verify cmd
This adds the cmd 'images verify', which verifies that previously built
and pushed containerdisks are bootable and that guests are working.

The example definitions of the docs are used to test containerdisks by
creating VMs with the respective containerdisk set as the boot source.
After waiting for a VM to become ready, a defined set of tests is run on
the corresponding VMI. If all tests succeed, the verified value of the
containerdisk is set to 'true' in the results file.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>