Add images verify cmd (#18)
* Prevent null pointer access when checking for unknown tag

This prevents a segfault when an unexpected error occurs, which cannot
be unwrapped (e.g. tls verification error on insecure registry).

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Improve logging of dry-runs

Do not always log 'Pushing' and let the user know what is really
happening instead.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Fix typos in log messages

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Move worker logic for image cmds into separate file

This moves the worker logic for image cmds into a separate file, so it
can be used in multiple image cmds. The logic was also refactored, so it
is easier to use, supports contexts and duplicated code during use is prevented.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Use const for shasum label

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Move PushImage to repository package

This moves the PushImage function into the repository package to make it
available for use in other image cmds. Also use the uniform identifier
'imgRef' for images.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Replace LayerFromReader with LayerFromOpener

The function tarball.LayerFromReader internally uses iotuil.ReadAll, which
buffers all data in memory and can lead to problems with large files.
Replace it with tarball.LayerFromOpener and implement a StreamLayerOpener,
that does not return an error chan and can be opened multiple times.

Thanks to akrejcir for the idea and proposal.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Overhaul creation of example VM definitions

The methods for creating example VM definitons are added to the artifact
interface so correct definitions for each type of guest can be created.
This also allows to test the example definitions in another command
(e.g. images verify).

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Do not add duplicate tag to AdditionalUniqueTags of CentOS Stream

The removed code added the main version of the CentOS Stream artifact to
AdditionalUniqueTags which is not an additional unique tag.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Refactor CLI flags

Flags that can be shared by image cmds are refactored into ImagesOptions
and moved to imagesCmd.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Write results file after pushing containerdisks

With this change the 'images push' command writes a results file after
building and pushing containerdisks. The results file contains all built
images, their sha256 sum and the pushed tags.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Add images verify cmd

This adds the cmd 'images verify', which verifies that previously built
and pushed containerdisks are bootable and that guests are working.

The example definitions of the docs are used to test containerdisks by
creating VMs with the respective containerdisk set as the boot source.
After waiting for a VM to become ready, a defined set of tests is run on
the corresponding VMI. If all tests succeed, the verified value of the
containerdisk is set to 'true' in the results file.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Add test to check if qemu-guest-agent is running

The test checks if the agent is running on the guest by attempting to
get the guest os info from the agent.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>

* Add test to check if ssh is working

The test checks if ssh is working on the guest by attempting to use the
credentials generated by images verify to login into the guest and to
run a simple command (e.g. 'echo hello').

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>