Pipeline for testing based on cirros (#24)
* Add a minimal pipeline example for testing

The pipeline brings up a cluster, force-builds the cirros containredisk,
verifies it and emulates a promotion.

Signed-off-by: Roman Mohr <rmohr@redhat.com>

* Ensure fast shutdown of VMIs

The greace period parameter on delete calls for VMIs does not work. It
is ignored on CRs and only works for pods. Setting a termination grace
period of 0 on the VMI to ensure swift shutdown.

Signed-off-by: Roman Mohr <rmohr@redhat.com>