fix: Add distro default usernames to artifacts
This adds the distro default usernames to the respective artifacts so
the examples in the generated docs are correct. Instead of potentially
adding the default user on distros using cloud-init again the docs
mention the default username in a comment now. On distros using ignition
providing a username is mandatory now.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>