fedora: Always sort releases before building artifact slices
As documented by the ArtifactsGatherer interface Gather() must return a
sorted list of gathered artifacts per release with the latest release
first. IOW and as tested currently this means it should return something
like the following:

[][]Artifact{
  {
    Artifact{release=40, arch=x86_64},
    Artifact{release=40, arch=aarch64},
  },
  {
    Artifact{release=39, arch=x86_64},
    Artifact{release=39, arch=aarach64},
  },
}

Maps are unordered [1] within golang so before we start processing we
need to extract the collected release versions and sort these to ensure
we get the desired order of Artifact slices.

[1] https://go.dev/ref/spec#Map_types

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>