fix: verify artifacts for the corresponding Kubernetes architecture
In the verification logic, the first artifact selected always
corresponded to the amd64 architecture. While this was not an
issue previously, it became apparent when attempting to enable
tests that require virtual machines with different configurations
for each architecture, such as Fedora.

This change is a necessary step for enabling the Fedora CI on
s390x. Additionally, it paves the way for refactoring the
configuration retrieval from common instance types.
Although this refactoring is outside the current scope, it can be
revisited in the near future.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>